# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ByteDock coding assessment platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  staff: 'staff',
  student: 'student',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatar?: string | null;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface TestCase {
  id: string;
  input: string;
  expectedOutput: string;
  isSample: boolean;
}

export interface TestCaseInput {
  input: string;
  expectedOutput: string;
  isSample?: boolean;
}

export type ProblemDifficulty = typeof ProblemDifficulty[keyof typeof ProblemDifficulty];


export const ProblemDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type ProblemStatus = typeof ProblemStatus[keyof typeof ProblemStatus];


export const ProblemStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export interface Problem {
  id: string;
  title: string;
  slug: string;
  difficulty: ProblemDifficulty;
  category: string;
  description: string;
  starterCode: string;
  status: ProblemStatus;
  acceptanceRate: number;
  solveCount: number;
  testCases: TestCase[];
  createdAt: string;
}

export type ProblemInputDifficulty = typeof ProblemInputDifficulty[keyof typeof ProblemInputDifficulty];


export const ProblemInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type ProblemInputStatus = typeof ProblemInputStatus[keyof typeof ProblemInputStatus];


export const ProblemInputStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export interface ProblemInput {
  /** @minLength 3 */
  title: string;
  category: string;
  difficulty: ProblemInputDifficulty;
  description: string;
  starterCode: string;
  status?: ProblemInputStatus;
}

export type ProblemUpdateDifficulty = typeof ProblemUpdateDifficulty[keyof typeof ProblemUpdateDifficulty];


export const ProblemUpdateDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type ProblemUpdateStatus = typeof ProblemUpdateStatus[keyof typeof ProblemUpdateStatus];


export const ProblemUpdateStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export interface ProblemUpdate {
  /** @minLength 3 */
  title?: string;
  category?: string;
  difficulty?: ProblemUpdateDifficulty;
  description?: string;
  starterCode?: string;
  status?: ProblemUpdateStatus;
}

export interface ExecuteInput {
  problemId: string;
  code: string;
}

export type ExecutionResultStatus = typeof ExecutionResultStatus[keyof typeof ExecutionResultStatus];


export const ExecutionResultStatus = {
  passed: 'passed',
  failed: 'failed',
  error: 'error',
  timeout: 'timeout',
} as const;

export interface TestResult {
  testCaseId: string;
  passed: boolean;
  actual: string;
  expected: string;
}

export interface ExecutionResult {
  status: ExecutionResultStatus;
  passed: number;
  total: number;
  runtimeMs: number;
  output: string;
  testResults: TestResult[];
}

export interface SubmissionInput {
  problemId: string;
  code: string;
}

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  accepted: 'accepted',
  rejected: 'rejected',
  error: 'error',
} as const;

export interface Submission {
  id: string;
  problemId: string;
  problemTitle: string;
  studentId: string;
  studentName: string;
  code: string;
  status: SubmissionStatus;
  passedTests: number;
  totalTests: number;
  runtimeMs: number;
  createdAt: string;
}

export interface StudentSummary {
  id: string;
  name: string;
  email: string;
  solvedCount: number;
  submissionCount: number;
  /** @nullable */
  lastActiveAt: string | null;
}

export interface Metric {
  label: string;
  value: number;
}

export interface StudentAnalytics {
  student: StudentSummary;
  accepted: number;
  rejected: number;
  byDifficulty: Metric[];
  recentSubmissions: Submission[];
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  submission: 'submission',
  solve: 'solve',
  publish: 'publish',
  register: 'register',
} as const;

export interface Activity {
  id: string;
  type: ActivityType;
  message: string;
  createdAt: string;
  userName: string;
}

export interface Progress {
  solvedCount: number;
  totalPublished: number;
  completionPercent: number;
  streak: number;
  byDifficulty: Metric[];
  recentActivity: Activity[];
}

export type DashboardSummaryRole = typeof DashboardSummaryRole[keyof typeof DashboardSummaryRole];


export const DashboardSummaryRole = {
  admin: 'admin',
  staff: 'staff',
  student: 'student',
} as const;

export interface DashboardSummary {
  role: DashboardSummaryRole;
  totalStudents: number;
  totalStaff: number;
  totalProblems: number;
  publishedProblems: number;
  totalSubmissions: number;
  acceptedSubmissions: number;
  activeStudents: number;
  solvedCount: number;
  completionPercent: number;
  recentActivity: Activity[];
  weeklySubmissions: Metric[];
}

/**
 * Error response
 */
export type ErrorResponse = Error;

export type ListProblemsParams = {
search?: string;
difficulty?: ListProblemsDifficulty;
status?: ListProblemsStatus;
};

export type ListProblemsDifficulty = typeof ListProblemsDifficulty[keyof typeof ListProblemsDifficulty];


export const ListProblemsDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type ListProblemsStatus = typeof ListProblemsStatus[keyof typeof ListProblemsStatus];


export const ListProblemsStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export type ListSubmissionsParams = {
problemId?: string;
studentId?: string;
};